fix(HTD-90): move cta to a custom hook
