+++ src/components/sections/pricing/plans/index.tsx
@@ -1,14 +1,12 @@
 'use client';
-import { upperFirst } from 'lodash';
 import { observer } from 'mobx-react-lite';
-import { usePostHog } from 'posthog-js/react';
 import { useCallback, useState } from 'react';
 
 import { PricingCard } from './components/card';
-import { StyledPricingCardCTAWrapper } from './components/card/card.styles';
 import { LoginInfo } from './components/login-info';
 import { Switch } from './components/switch';
 import type { SwitchProps } from './components/switch/switch.types';
+import { usePlanCta } from './hooks/get-plan-cta';
 import {
   StyledPricingPlansCardsWrapper,
   StyledPricingPlansDisclaimerWrapper,
@@ -18,10 +16,8 @@ import {
 } from './plans.styles';
 import type { PricingPlansData, PricingPlansProps } from './plans.types';
 
-import { Button } from '@/components/elements/button';
-import { PaperPlaneTilt, Sparkle, Spinner } from '@/components/elements/icon';
+import { Spinner } from '@/components/elements/icon';
 import { Text } from '@/components/elements/text';
-import { DownloadButton } from '@/components/modules/download-button';
 import { AccountStore } from '@/lib/store/account-store';
 
 const pricingPlans: SwitchProps['options'] = [
@@ -107,8 +103,7 @@ const { cards, disclaimer }: PricingPlansData = {
 
 export const PricingPlans = observer(({}: PricingPlansProps) => {
   const [account] = useState(() => new AccountStore());
-  const posthog = usePostHog();
-
+  const getPlanCTA = usePlanCta();
   const [planCycle, setPlanCycle] = useState(pricingPlans[0].id);
   const isAnnual = planCycle === 'annual';
 
@@ -141,61 +136,6 @@ export const PricingPlans = observer(({}: PricingPlansProps) => {
     return paidCycle === planCycle ? statusDescription : `${statusDescription} (${paidCycle})`;
   }, []);
 
-  const getPlanCta = useCallback((tierCode: string) => {
-    const { paidTier, paidCycle } = account.subscription;
-
-    if (waitingForPurchase) {
-      return <Button icon={Spinner}></Button>;
-    }
-
-    if (tierCode === 'free') {
-      return <DownloadButton />;
-    }
-
-    if (paidTier === tierCode) {
-      if (paidCycle === planCycle) {
-        return (
-          <StyledPricingCardCTAWrapper>
-            <Text fontSize="s" color="lightGrey">
-              Download now and log in to access your {upperFirst(tierCode)} subscription
-            </Text>
-            <DownloadButton />
-          </StyledPricingCardCTAWrapper>
-        );
-      }
-
-      return (
-        <StyledPricingCardCTAWrapper>
-          <Text fontSize="s" color="lightGrey">
-            You already have this {paidCycle} plan.
-          </Text>
-          <Button href="/contact/">Change to {planCycle}</Button>
-        </StyledPricingCardCTAWrapper>
-      );
-    }
-
-    if (tierCode === 'pro') {
-      return (
-        <Button icon={Sparkle} onClick={() => account.buyPlan('pro', planCycle, posthog)}>
-          Buy Pro
-        </Button>
-      );
-    }
-
-    if (tierCode === 'team') {
-      return (
-        <Button
-          $variant="secondary"
-          href="/contact/"
-          icon={PaperPlaneTilt}
-          onClick={() => account.reportPurchaseEvent('Select team plan', tierCode, planCycle, posthog)}
-        >
-          Get in touch
-        </Button>
-      );
-    }
-  }, []);
-
   return (
     <>
       <StyledPricingPlansWrapper>
@@ -219,7 +159,7 @@ export const PricingPlans = observer(({}: PricingPlansProps) => {
                 price={getPlanMonthlyPrice(card.id)}
                 {...card}
               >
-                {getPlanCta(card.id)}
+                {getPlanCTA(card.id, account, waitingForPurchase, planCycle)}
               </PricingCard>
             ))}
         </StyledPricingPlansCardsWrapper>
